services/edge_api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
pydantic>=2.7.0
guardrails-ai>=0.5.0
